# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * SRS Invoice Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface Client {
  id: number;
  name: string;
  address: string;
  /** @nullable */
  gstin?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  createdAt?: string;
}

export interface ClientInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  address: string;
  gstin?: string;
  phone?: string;
  email?: string;
}

export interface ClientUpdate {
  /** @minLength 1 */
  name?: string;
  address?: string;
  gstin?: string;
  phone?: string;
  email?: string;
}

export interface Panel {
  id: number;
  name: string;
  breakdownText: string;
  panelSize: string;
  price: number;
  defaultQty: number;
  createdAt?: string;
}

export interface PanelInput {
  /** @minLength 1 */
  name: string;
  breakdownText?: string;
  panelSize?: string;
  /** @minimum 0 */
  price?: number;
  /** @minimum 1 */
  defaultQty?: number;
}

export interface PanelUpdate {
  /** @minLength 1 */
  name?: string;
  breakdownText?: string;
  panelSize?: string;
  /** @minimum 0 */
  price?: number;
  /** @minimum 1 */
  defaultQty?: number;
}

export interface TechSpecItem {
  id: number;
  itemName: string;
  defaultSpec: string;
  createdAt?: string;
}

export interface TechSpecItemInput {
  /** @minLength 1 */
  itemName: string;
  /** @minLength 1 */
  defaultSpec: string;
}

export interface TechSpecItemUpdate {
  /** @minLength 1 */
  itemName?: string;
  /** @minLength 1 */
  defaultSpec?: string;
}

export interface Category {
  id: number;
  name: string;
  createdAt?: string;
}

export interface CategoryInput {
  /** @minLength 1 */
  name: string;
}

export interface CategoryUpdate {
  /** @minLength 1 */
  name?: string;
}

export interface SubCategory {
  id: number;
  categoryId: number;
  name: string;
  createdAt?: string;
}

export interface SubCategoryInput {
  categoryId: number;
  /** @minLength 1 */
  name: string;
}

export interface SubCategoryUpdate {
  categoryId?: number;
  /** @minLength 1 */
  name?: string;
}

export interface Item {
  id: number;
  subCategoryId: number;
  name: string;
  /** @nullable */
  catNo?: string | null;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  packQty?: number | null;
  /** @nullable */
  specifications?: string | null;
  createdAt?: string;
}

export interface ItemInput {
  subCategoryId: number;
  /** @minLength 1 */
  name: string;
  catNo?: string;
  /** @minimum 0 */
  price?: number;
  /** @minimum 0 */
  packQty?: number;
  specifications?: string;
}

export interface ItemUpdate {
  subCategoryId?: number;
  /** @minLength 1 */
  name?: string;
  catNo?: string;
  /** @minimum 0 */
  price?: number;
  /** @minimum 0 */
  packQty?: number;
  specifications?: string;
}

export interface InvoiceItem {
  id: number;
  invoiceId: number;
  sNo: number;
  description: string;
  /** @nullable */
  hsnSac?: string | null;
  /** @nullable */
  qty?: number | null;
  /** @nullable */
  rate?: number | null;
  /** @nullable */
  per?: string | null;
  amount: number;
}

export interface InvoiceItemInput {
  sNo: number;
  /** @minLength 1 */
  description: string;
  hsnSac?: string;
  qty?: number;
  rate?: number;
  per?: string;
  amount: number;
}

export interface Invoice {
  id: number;
  invoiceNo: number;
  date: string;
  clientId: number;
  client: Client;
  /** @nullable */
  workSite?: string | null;
  /** @nullable */
  deliveryNote?: string | null;
  /** @nullable */
  modeOfPayment?: string | null;
  /** @nullable */
  suppliersRef?: string | null;
  /** @nullable */
  othersRef?: string | null;
  /** @nullable */
  buyersOrderNo?: string | null;
  /** @nullable */
  buyersOrderDate?: string | null;
  /** @nullable */
  despatchDocNo?: string | null;
  /** @nullable */
  despatchDocDate?: string | null;
  /** @nullable */
  despatchedThrough?: string | null;
  /** @nullable */
  destination?: string | null;
  /** @nullable */
  termsOfDelivery?: string | null;
  items: InvoiceItem[];
  subtotal: number;
  cgstRate: number;
  sgstRate: number;
  cgstAmount: number;
  sgstAmount: number;
  roundOff: number;
  netTotal: number;
  amountInWords: string;
  createdAt: string;
}

export interface InvoiceInput {
  invoiceNo: number;
  date: string;
  clientId: number;
  workSite?: string;
  deliveryNote?: string;
  modeOfPayment?: string;
  suppliersRef?: string;
  othersRef?: string;
  buyersOrderNo?: string;
  buyersOrderDate?: string;
  despatchDocNo?: string;
  despatchDocDate?: string;
  despatchedThrough?: string;
  destination?: string;
  termsOfDelivery?: string;
  items: InvoiceItemInput[];
  cgstRate: number;
  sgstRate: number;
}

export interface InvoiceUpdate {
  date?: string;
  clientId?: number;
  workSite?: string;
  deliveryNote?: string;
  modeOfPayment?: string;
  suppliersRef?: string;
  othersRef?: string;
  buyersOrderNo?: string;
  buyersOrderDate?: string;
  despatchDocNo?: string;
  despatchDocDate?: string;
  despatchedThrough?: string;
  destination?: string;
  termsOfDelivery?: string;
  items?: InvoiceItemInput[];
  cgstRate?: number;
  sgstRate?: number;
}

export interface InvoiceSummary {
  id: number;
  invoiceNo: number;
  date: string;
  clientId: number;
  clientName: string;
  /** @nullable */
  workSite?: string | null;
  netTotal: number;
  createdAt: string;
}

export interface InvoiceStats {
  totalInvoices: number;
  totalRevenue: number;
  thisMonthRevenue: number;
  thisMonthCount: number;
  lastMonthRevenue: number;
  lastMonthCount: number;
}

export interface NextInvoiceNumber {
  nextNumber: number;
}

export type ListSubCategoriesParams = {
categoryId?: number;
};

export type ListItemsParams = {
subCategoryId?: number;
};

export type ListInvoicesParams = {
clientId?: number;
search?: string;
};