# lib/api-zod/src/generated/types/item.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * SRS Invoice Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface Item {
  id: number;
  subCategoryId: number;
  name: string;
  /** @nullable */
  catNo?: string | null;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  packQty?: number | null;
  /** @nullable */
  specifications?: string | null;
  createdAt?: string;
}