# lib/api-zod/src/generated/types/itemUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * SRS Invoice Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface ItemUpdate {
  subCategoryId?: number;
  /** @minLength 1 */
  name?: string;
  catNo?: string;
  /** @minimum 0 */
  price?: number;
  /** @minimum 0 */
  packQty?: number;
  specifications?: string;
}