# lib/api-zod/src/generated/types/subCategory.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * SRS Invoice Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface SubCategory {
  id: number;
  categoryId: number;
  name: string;
  createdAt?: string;
}