# lib/api-zod/src/generated/types/subCategoryInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * SRS Invoice Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface SubCategoryInput {
  categoryId: number;
  /** @minLength 1 */
  name: string;
}